Video_Feature_Attack_4060/requirements_4060.txt
torch
transformers==4.43.1
numpy
opencv-python
Pillow


